refactor: abstract chat related code into its own file
